feat: add ccrwrapper endpoints
